refactor(fetcher): move FetchTimeoutError to timeout.ts
- Relocate FetchTimeoutError class from fetcherError.ts to timeout.ts
- Update related test files to import FetchTimeoutError from new location
- Remove FetchTimeoutError tests from fetcherError.comprehensive.test.ts
- Add FetchTimeoutError tests to timeout.test.ts
