+++ packages/fetcher/test/timeout.test.ts
@@ -13,7 +13,15 @@
 
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { FetchTimeoutError, HttpMethod, resolveTimeout, TimeoutCapable, timeoutFetch } from '../src';
+import {
+  FetcherError,
+  FetchRequest,
+  FetchTimeoutError,
+  HttpMethod,
+  resolveTimeout,
+  TimeoutCapable,
+  timeoutFetch,
+} from '../src';
 
 describe('timeout.ts', () => {
   describe('resolveTimeout', () => {
@@ -251,5 +259,56 @@ describe('timeout.ts', () => {
       // Clean up
       mockFetch.mockRestore();
     });
+  });
+});
+
+describe('FetchTimeoutError', () => {
+  const mockRequest = {
+    url: '/test/timeout',
+    method: 'GET',
+    timeout: 5000,
+  } as FetchRequest;
+
+  it('should create FetchTimeoutError with request', () => {
+    const timeoutError = new FetchTimeoutError(mockRequest);
+
+    expect(timeoutError).toBeInstanceOf(FetchTimeoutError);
+    expect(timeoutError).toBeInstanceOf(FetcherError);
+    expect(timeoutError).toBeInstanceOf(Error);
+    expect(timeoutError.name).toBe('FetchTimeoutError');
+    expect(timeoutError.message).toBe(
+      `Request timeout of ${mockRequest.timeout}ms exceeded for ${mockRequest.method} ${mockRequest.url}`,
+    );
+    expect(timeoutError.request).toBe(mockRequest);
+    expect(timeoutError.cause).toBeUndefined();
+  });
+
+  it('should create FetchTimeoutError with different request parameters', () => {
+    const differentRequest = {
+      url: 'https://api.example.com/users',
+      method: 'POST',
+      timeout: 1000,
+    } as FetchRequest;
+    const timeoutError = new FetchTimeoutError(differentRequest);
+
+    expect(timeoutError).toBeInstanceOf(FetchTimeoutError);
+    expect(timeoutError.message).toBe(
+      `Request timeout of ${differentRequest.timeout}ms exceeded for ${differentRequest.method} ${differentRequest.url}`,
+    );
+    expect(timeoutError.request).toBe(differentRequest);
+  });
+
+  it('should use default method when not specified', () => {
+    const requestWithoutMethod = {
+      url: '/test/default-method',
+      timeout: 3000,
+    } as FetchRequest;
+    const timeoutError = new FetchTimeoutError(requestWithoutMethod);
+
+    expect(timeoutError).toBeInstanceOf(FetchTimeoutError);
+    expect(timeoutError.message).toBe(
+      `Request timeout of ${requestWithoutMethod.timeout}ms exceeded for GET ${requestWithoutMethod.url}`,
+    );
+    expect(timeoutError.request).toBe(requestWithoutMethod);
   });
 });
